move dimension table stuff to create tables instead, only run in the beginning
